Refactoring, remove images.json from download API, updated hook callback with images database
