+++ libs/TaskManager.js
@@ -16,18 +16,17 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 "use strict";
-let assert = require('assert');
-let config = require('../config');
-let rmdir = require('rimraf');
-let fs = require('fs');
-let path = require('path');
-let logger = require('./logger');
-let Task = require('./Task');
-let statusCodes = require('./statusCodes');
-let async = require('async');
-let schedule = require('node-schedule');
-let Directories = require('./Directories');
-let request = require('request');
+const assert = require('assert');
+const config = require('../config');
+const rmdir = require('rimraf');
+const fs = require('fs');
+const path = require('path');
+const logger = require('./logger');
+const Task = require('./Task');
+const statusCodes = require('./statusCodes');
+const async = require('async');
+const schedule = require('node-schedule');
+const Directories = require('./Directories');
 
 const TASKS_DUMP_FILE = path.join(Directories.data, "tasks.json");
 const CLEANUP_TASKS_IF_OLDER_THAN = 1000 * 60 * config.cleanupTasksAfter; // minutes
@@ -150,34 +149,8 @@ module.exports = class TaskManager{
 			if (task){
 				this.addToRunningQueue(task);
 				task.start(() => {
-					// Hooks can be passed via command line 
-					// or for each individual task
-					const hooks = [task.webhook, config.webhook];
 
-					hooks.forEach(hook => {
-						if (hook && hook.length > 3){
-							const notifyCallback = (attempt) => {
-								if (attempt > 5){
-									logger.warn(`Callback failed, will not retry: ${hook}`);
-									return;
-								}
-								request.post(hook, {
-										json: task.getInfo()
-									},
-									(error, response) => {
-										if (error || response.statusCode != 200){
-											logger.warn(`Callback failed, will retry in a bit: ${hook}`);
-											setTimeout(() => {
-												notifyCallback(attempt + 1);
-											}, attempt * 5000);
-										}else{
-											logger.debug(`Callback invoked: ${hook}`);
-										}
-								});
-							};
-							notifyCallback(0);
-						}
-					});
+					task.callWebhooks();
 
 					this.removeFromRunningQueue(task);
 					this.processNextTask();
